Inline the projection module
It's now just a single function.
