+++ src/persist-client/src/lib.rs
@@ -84,7 +84,6 @@ pub mod operators {
         operator before yielding.",
     );
 }
-pub mod project;
 pub mod read;
 pub mod rpc;
 pub mod schema;
